# docs/openapi/api.v1.json
{
  "components" : {
    "parameters" : {
      "CorrelationId" : {
        "description" : "Optional trace ID; generated if absent and returned on every response",
        "in" : "header",
        "name" : "X-Correlation-Id",
        "required" : false,
        "schema" : {
          "example" : "demo-123",
          "type" : "string"
        }
      }
    },
    "schemas" : {
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "correlationId" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "violations" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : "array"
          }
        }
      },
...
      "FieldViolation" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "LoginRequest" : {
        "properties" : {
          "password" : {
            "example" : "Password1!",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "username" : {
            "example" : "adjuster1",
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "password", "username" ],
        "type" : "object"
      },
...
      "NotificationView" : {
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "claimId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "readAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageResponseNotificationView" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/NotificationView"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RefreshRequest" : {
        "properties" : {
          "refreshToken" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "TokenResponse" : {
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "accessTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "expiresIn" : {
            "description" : "Access token lifetime in seconds",
            "example" : 900,
            "format" : "int64",
            "type" : "integer"
          },
          "refreshToken" : {
            "description" : "Single-use: every refresh returns a new one and revokes the old one",
            "type" : "string"
          },
          "refreshTokenExpiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "tokenType" : {
            "example" : "Bearer",
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserSummary"
          }
        },
        "type" : "object"
      },
      "UnreadCount" : {
        "properties" : {
          "unread" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UserSummary" : {
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "role" : {
            "enum" : [ "CLAIMANT", "ADJUSTER", "SUPERVISOR", "SIU" ],
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login (valid 15 minutes)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Guidewire-style claims handling: FNOL, exposures, reserves, payments, approvals, SIU and AI document assessment.\n\nAPI version **v1** (contract 1.3.0). Breaking changes get `/api/v2`; additions bump the minor version.",
    "title" : "AI Claims Platform API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/auth/login" : {
      "post" : {
        "description" : "Demo users: claimant1, claimant2, adjuster1, adjuster2, supervisor1, siu1 (password Password1!)",
        "operationId" : "login",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too many requests"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "security" : [ ],
        "summary" : "Log in",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "description" : "Always 204, even for an unknown token.",
        "operationId" : "logout",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Revoked (or was already unknown or revoked)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "security" : [ ],
        "summary" : "Revoke a refresh token",
        "tags" : [ "Auth" ]
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "description" : "The refresh token is single-use. Reusing an old one ends all of the user's sessions.",
        "operationId" : "refreshTokens",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TokenResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "security" : [ ],
        "summary" : "Exchange a refresh token for new tokens",
        "tags" : [ "Auth" ]
      }
    },
...
    "/api/v1/portal/notifications" : {
      "get" : {
        "operationId" : "listMyNotifications",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseNotificationView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "My notifications, newest first",
        "tags" : [ "Portal: notifications" ]
      }
    },
    "/api/v1/portal/notifications/unread-count" : {
      "get" : {
        "operationId" : "getMyUnreadNotificationCount",
        "parameters" : [ {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UnreadCount"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Number of unread notifications (for the badge)",
        "tags" : [ "Portal: notifications" ]
      }
    },
    "/api/v1/portal/notifications/{id}/read" : {
      "post" : {
        "operationId" : "markNotificationRead",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NotificationView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Mark as read (idempotent)",
        "tags" : [ "Portal: notifications" ]
      }
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8081"
  } ],
...
}